+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Address.hs
@@ -10,7 +10,6 @@ import Generics.SOP qualified as SOP
 import Plutarch.LedgerApi.Utils (PMaybeData)
 import Plutarch.LedgerApi.V1.Credential (PCredential, PStakingCredential)
 import Plutarch.Prelude
-import Plutarch.Repr.Data (DeriveAsDataStruct (DeriveAsDataStruct))
 import PlutusLedgerApi.V1 qualified as Plutus
 
 -- | @since 2.0.0